Moved files and added namespace.
