+++ slider/src/output/displayBuffer.cpp
@@ -1,19 +1,19 @@
 #include "displayBuffer.h"
 
-DisplayBuffer::DisplayBuffer():
+Output::DisplayBuffer::DisplayBuffer():
     m_Cursor(0), m_Buffer()
 {
     m_Buffer.fill(' ');
 }
 
-void DisplayBuffer::fillCurrentLine()
+void Output::DisplayBuffer::fillCurrentLine()
 {
     const auto maxCursor = ((m_Cursor / LCD_LINE_LENGTH) + 1) * LCD_LINE_LENGTH;
     while (m_Cursor < maxCursor)
         m_Buffer[m_Cursor++] = ' ';
 }
 
-size_t DisplayBuffer::write(Keycode value)
+size_t Output::DisplayBuffer::write(Keycode value)
 {
     //do not go over the line!
     const auto maxCursor = ((m_Cursor / LCD_LINE_LENGTH) + 1) * LCD_LINE_LENGTH;
@@ -22,7 +22,7 @@ size_t DisplayBuffer::write(Keycode value)
     return 1;
 }
 
-void DisplayBuffer::SetCursor(const int line, const int column)
+void Output::DisplayBuffer::SetCursor(const int line, const int column)
 {
     if (line >= LCD_NUM_LINES || line < 0)
         return;
